+++ packages/spotlight/CHANGELOG.md
@@ -4,6 +4,10 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ## [unreleased]
 
+### Added
+
+- `spotlight.spottableClass` export
+
 ### Fixed
 
 - `spotlight.Spotlight` behavior to follow container config rules when navigating between containers